requirements.txt
pillow
opencv-python
numpy
dearpygui
scikit-image
pytest
